# docs/openapi/openapi-drapt.yaml
openapi: 3.1.0    # Sửa lại 
info:
  title: Health Tracking API
  version: 0.1.0
  description: >
    Draft OpenAPI contract aligned with the project's flowcharts and database models.
    Authentication: Bearer access token. Most write/read endpoints require auth.
servers:
  - url: /api/v1
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user + profile
      operationId: users_me
      security:
        - HTTPBearer: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/dashboard:
    get:
      tags: [Admin]
      summary: Admin dashboard summary
      operationId: admin_dashboard
      security:
        - HTTPBearer: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboardResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
    UserRole:
      type: string
      enum: [user, admin]
    Gender:
      type: string
      enum: [male, female]
...
    UserPublic:
      type: object
      required: [id, username, email, role]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
    Profile:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          format: uuid
        full_name:
          type: string
          nullable: true
        gender:
          $ref: '#/components/schemas/Gender'
        date_of_birth:
          type: string
          format: date
          nullable: true
        height_cm_default:
          type: number
          nullable: true
...
    UserMeResponse:
      type: object
      required: [user, profile]
      properties:
        user:
          $ref: '#/components/schemas/UserPublic'
        profile:
          $ref: '#/components/schemas/Profile'
...
    AdminDashboardResponse:
      type: object
      required: [users_count, posts_count, open_tickets_count]
      properties:
        users_count:
          type: integer
          minimum: 0
        posts_count:
          type: integer
          minimum: 0
        open_tickets_count:
          type: integer
          minimum: 0
...